+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - KeyboardShortcuts.copy() added.
 ### is.codion.swing.framework.ui
 - EntityPanel, EntityTablePanel, keyboard shortcuts can now be configured on an instance basis, via configure().keyStroke().
+- TabbedPanelLayout.Builder.keyboardShortcut() added for configuring keyboard shortcuts on an instance basis.
 
 ## 0.17.22
 ### is.codion.swing.common.core